+++ rust/candid_parser/Cargo.toml
@@ -18,7 +18,7 @@ build = "build.rs"
 lalrpop = "0.20.0"
 
 [dependencies]
-candid = { path = "../candid", version = "0.10" }
+candid = { path = "../candid", version = "0.10", features = ["all"] }
 codespan-reporting = "0.11"
 hex.workspace = true
 num-bigint.workspace = true
